git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@3214 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkSim/src/rwsim/control/BeamJointController.cpp
@@ -60,7 +60,7 @@ void BeamJointController::update(double dt, rw::kinematics::State& state) {
 	// the error in configuration result in a torque
 	Q torque = q;
 
-	for(int i=0;i<q.size();i++){ torque[i] = 2; }
+	for(size_t i=0;i<q.size();i++){ torque[i] = 2; }
 
 	_ddev->setForceLimit(torque);
 	_ddev->setVelocity(q_error, state);
